Move user service to services package
