+++ h/accounts/__init__.py
@@ -52,8 +52,6 @@ def includeme(config):
     # `User` instance.
     config.add_request_method(authenticated_user, property=True)
 
-    config.register_service_factory('.services.user_service_factory',
-                                    name='user')
     config.register_service_factory('.services.user_signup_service_factory',
                                     name='user_signup')
