+++ tests/h/views/activity_test.py
@@ -736,7 +736,7 @@ class TestUserSearchController(object):
 
     @pytest.fixture
     def user_service(self, patch, pyramid_config):
-        user_service = patch('h.accounts.services.UserService')
+        user_service = patch('h.services.user.UserService')
         pyramid_config.register_service(user_service, name='user')
         return user_service
